Move webChannel simulation functions into shared file.
This means that these functions will be usable from UrlManager.test.js.
I'll make use of that in a future PR.
